feat(auth): add refresh token

+++ internal/models/login.go
@@ -13,7 +13,7 @@ func (l LoginRequest) Validate() error {
 }
 
 type LoginResponse struct {
-	UserID       uint64 `json:"user_id"`
+	UserID       int    `json:"user_id"`
 	Username     string `json:"username"`
 	Fullname     string `json:"full_name"`
 	Email        string `json:"email"`
